covid_stretch/requirements.txt
appdirs>=1.4.4
asgiref>=3.3.1
astroid>=2.4.2
attrs>=20.3.0
bandit>=1.6.2
beautifulsoup4>=4.9.3
# black>=20.8b1
certifi>=2020.11.8
chardet>=3.0.4
click>=7.1.2
coverage>=5.3
Django>=3.2.10,<4.0
django-cors-headers>=3.5.0
django-crispy-forms>=1.12.0
django-extensions>=3.0.9
django-multiforloop>=0.2.1
# django-multipleselectfield>=0.2.0.post2
django-multiselectfield>=0.1.12
django-phone-field>=1.8.1
django-widget-tweaks>=1.4.8
djangorestframework>=3.12.2
dodgy>=0.2.1
factory-boy>=3.1.0
Faker>=12.3.0
flake8>=3.8.4
flake8-polyfill>=1.0.2
fontawesomefree>=6.0.0
freezegun>=1.0.0
gitdb>=4.0.5
GitPython>=3.1.11
gunicorn>=20.1.0
idna>=2.10
importlib-metadata>=2.0.0
isort>=5.6.4
lazy-object-proxy>=1.4
lxml>=4.6.1
mccabe>=0.6.1
more-itertools>=8.6.0
mypy-extensions>=0.4.3
numpy>=1.19.4
packaging>=20.4
pandas>=1.1.4
pathspec>=0.8.1
pbr>=5.5.1
pep8-naming>=0.11.1
Pillow>=8.0.1
pycodestyle>=2.6.0
pyflakes>=2.2.0
pyparsing>=2.4.7
python-dateutil>=2.8.1
pytz>=2020.4
PyYAML>=5.3.1
regex>=2020.11.13
simple-term-menu>=1.4.1
six>=1.15.0
smmap>=3.0.4
soupsieve>=2.0.1
sqlparse>=0.4.1
stevedore>=3.2.2
text-unidecode>=1.3
toml>=0.10.2
typed-ast>=1.4.1
typing-extensions>=3.7.4.3
whitenoise>=6.0.0
wrapt>=1.12.1
zipp>=3.4.0
